server/ufw: test düzelt

+++ lib/scripts/server/firewall_test.sh
@@ -2,25 +2,37 @@
 
 set -euo pipefail; [[ -z ${TRACE:-} ]] || set -x
 
-systemctl is-enabled ufw && false
+if [[ -z ${firewall_enable:-} ]]; then
+	systemctl is-enabled ufw && false
 
-systemctl enable ufw && systemctl start ufw
+	systemctl enable ufw && systemctl start ufw
 
-goss -g - validate --format documentation <<-EOF
-	package:
-	  ufw:
-	    installed: true
-	service:
-	  ufw:
-	    enabled: true
-	    running: true
-EOF
+	goss -g - validate --format documentation <<-EOF
+		package:
+		  ufw:
+		    installed: true
+		service:
+		  ufw:
+		    enabled: true
+		    running: true
+	EOF
 
-systemctl stop ufw && systemctl disable ufw
+	systemctl stop ufw && systemctl disable ufw
 
-goss -g - validate --format documentation <<-EOF
-	service:
-	  ufw:
-	    enabled: false
-	    running: false
-EOF
+	goss -g - validate --format documentation <<-EOF
+		service:
+		  ufw:
+		    enabled: false
+		    running: false
+	EOF
+else
+	goss -g - validate --format documentation <<-EOF
+		package:
+		  ufw:
+		    installed: true
+		service:
+		  ufw:
+		    enabled: true
+		    running: true
+	EOF
+fi
